Moving w_auth_hash_key method to another source code file
