+++ src/headers/sec.h
@@ -14,6 +14,7 @@
 #include <time.h>
 #include <pthread.h>
 #include "shared.h"
+#include "os_crypto/sha1/sha1_op.h"
 
 typedef enum _crypt_method{
     W_METH_BLOWFISH,W_METH_AES
@@ -189,6 +190,15 @@ int OS_DeleteSocket(keystore * keys, int sock);
  * @retval REMOTED_NET_PROTOCOL_UDP if agent protocol is UDP
  */
 int w_get_agent_net_protocol_from_keystore(keystore * keys, const char * agent_id);
+
+/**
+ * @brief Receives a keyentry structure and returns the SHA1 value of the agent's key.
+ *
+ * @param key_entry The agent's key structure
+ * @param output The variable where the hashed key will be stored
+ * @retval Returns OS_SUCCESS on success or OS_INVALID on error.
+ **/
+int w_auth_hash_key(keyentry *key_entry, os_sha1 output);
 
 /* Set the agent crypto method read from the ossec.conf file */
 void os_set_agent_crypto_method(keystore * keys,const int method);
