+++ src/os_crypto/shared/keys.c
@@ -763,3 +763,19 @@ int OS_WriteTimestamps(keystore * keys) {
 
     return r;
 }
+int w_auth_hash_key(keyentry *key_entry, os_sha1 output) {
+  if (!key_entry || !output) {
+    mdebug2("Unable to hash agent's key due to empty parameters.");
+    return OS_INVALID;
+  }
+
+  char *key = key_entry->key;
+
+  if (key) {
+    return OS_SHA1_Str(key, strlen(key), output);
+  }
+  else {
+    mdebug2("Unable to hash agent's key due to empty value.");
+    return OS_INVALID;
+  }
+}
